pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel", "build"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app", "frontend"]

[tool.setuptools.dynamic-dependencies]
dependencies = []

[project]
name = "trending-papers-api"
version = "1.0.0"
description = "Trending Papers Discovery API"
authors = [
    {name = "MrBozkay", email = "mrbozkay@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "gradio==3.50.0",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "aiohttp==3.9.1",
    "asyncio-mqtt==0.16.1",
    "jinja2==3.1.2",
    "python-multipart==0.0.6",
    "arxiv==1.4.8",
    "PyGithub==1.59.1",
    "beautifulsoup4==4.12.2",
    "lxml==4.9.3",
    "redis==5.0.1",
    "tenacity==8.2.3",
    "slowapi==0.1.9",
    "httpx==0.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
trending-papers-api = "app.main:app"

[project.urls]
Homepage = "https://github.com/MrBozkay/trending-papers-api"
Repository = "https://github.com/MrBozkay/trending-papers-api"
"Bug Tracker" = "https://github.com/MrBozkay/trending-papers-api/issues"